Extract AuthenticationContext to separate file for fast refresh
Moves AuthenticationContext from authentication.container.tsx to its own
file so the container only exports components. Adds shouldRevalidate to
the eslint react-refresh allowExportNames list.
