+++ app/modules/authentication/authentication.context.ts
@@ -1,0 +1,4 @@
+import { createContext } from "react";
+import type { User } from "~/modules/users/users.types";
+
+export const AuthenticationContext = createContext<User | null>(null);
