+++ app/modules/projects/helpers/getProjectsItemActions.tsx
@@ -1,7 +1,7 @@
 import type { CollectionItemAction } from "@/components/ui/collectionContentItem";
 import { Edit, Trash2 } from "lucide-react";
 import { useContext } from "react";
-import { AuthenticationContext } from "~/modules/authentication/containers/authentication.container";
+import { AuthenticationContext } from "~/modules/authentication/authentication.context";
 import type { User } from "~/modules/users/users.types";
 import ProjectAuthorization from "../authorization";
 import type { Project } from "../projects.types";
