+++ app/modules/teams/hooks/useTeamAuthorization.ts
@@ -1,5 +1,5 @@
 import { useContext } from "react";
-import { AuthenticationContext } from "~/modules/authentication/containers/authentication.container";
+import { AuthenticationContext } from "~/modules/authentication/authentication.context";
 import type { User } from "~/modules/users/users.types";
 import TeamAuthorization, {
   type TeamAuthorizationShape,
